Upgraded Transcode script to use multiprocessing. Improved output to show unmatched files/tracks.

+++ Transcode.py
@@ -2,11 +2,14 @@ import os
 import os.path
 import argparse
 import re
-from mfile import open_music_file, mapping as mfile_mapping
+from concurrent.futures import ProcessPoolExecutor
+import operator
+from itertools import repeat
 
 # from mutagen.mp3 import MP3
 # from EasyID3Custom import EasyID3Custom as EasyID3
 
+from mfile import open_music_file, mapping as mfile_mapping
 import config
 from core.metadata import Metadata
 from core.util import get_fnames
@@ -19,22 +22,22 @@ from core.track import Track
 
 http_re = re.compile(r'^https?://', flags=re.IGNORECASE)
 
-def convert(infile, outfile, metadata, out_ext, bitrate):
-    # Sanity check for bitrate
-    if bitrate > 2000:
-        bitrate //= 1000
-    in_md = open_music_file(infile)
-    decoder = in_md.create_decoder()
+def convert(infile, outfile, metadata, out_ext, bitrate, test):
+    if not test:
+        # Sanity check for bitrate
+        if bitrate > 2000:
+            bitrate //= 1000
+        in_md = open_music_file(infile)
+        decoder = in_md.create_decoder()
 
-    if os.path.exists(outfile):
-        os.remove(outfile)
-    encoder = mfile_mapping[out_ext].create_encoder(outfile, metadata, bitrate)
+        if os.path.exists(outfile):
+            os.remove(outfile)
+        encoder = mfile_mapping[out_ext].create_encoder(outfile, metadata, bitrate)
 
-    decodedData, errs = decoder.communicate()
-    encoder.communicate(decodedData)
+        decodedData, errs = decoder.communicate()
+        encoder.communicate(decodedData)
 
-    encoded = open_music_file(outfile)
-    return encoded
+    return outfile
 
 def transcode(input_files, oom, bitrate, test):
 
@@ -42,52 +45,65 @@ def transcode(input_files, oom, bitrate, test):
 
     output_tracks = list()
     if os.path.isfile(oom) or http_re.match(oom):
-        output_tracks = [Track.from_metadata(m) for m in get_track_list(oom)]
+        output_tracks = [Track.from_metadata(m, match_to_existing=False) for m in get_track_list(oom)]
     else:
         output_tracks = [Track.from_file(fname, default_metadata='db') for fname in get_fnames(oom)]
 
     matched, unmatched_inputs, unmatched_outputs = match_metadata_to_files(input_files, output_tracks)
 
-    for input_track, output_track in matched:
-        fname = input_track.location
-        dest = getattr(output_track, 'location', output_track.calculate_fname())
+    matched_inputs, matched_outputs = list(zip(*matched))
+    sources = list(map(operator.attrgetter('location'), matched_inputs))
+    dests = [getattr(o, 'location', o.calculate_fname()) for o in matched_outputs]
+    output_dir = os.path.dirname(dests[0])
+    if not os.path.isdir(output_dir):
+        os.makedirs(output_dir)
+    exts = [os.path.splitext(dest)[1] for dest in dests]
 
-        if not test:
-            ext = os.path.splitext(dest)[1]
-            encoded = convert(fname, dest, output_track, ext, bitrate)
-        else:
-            encoded = Metadata.from_dict({'bitrate': bitrate * 1000,
-                                          'fsize': int(input_track.length * bitrate / 8)})
+    with ProcessPoolExecutor() as executor:
+        for (input_track, output_track), fname, dest, _ in \
+            zip(matched, sources, dests, executor.map(convert, sources, dests, matched_outputs,
+                                                      exts, repeat(bitrate), repeat(test))):
 
-        if output_track.db is not None:
-            output_track.db.update(encoded, False)
-
-        # Fancy string formatting
-        stripped_fname, fname_ext = os.path.splitext(fname)
-        stripped_dest, dest_ext = os.path.splitext(dest)
-        max_fbase_length = max(len(stripped_fname), len(stripped_dest))
-
-        print('Transcoded %s%s\n        to %s%s with the metadata\n%s' %
-                            (stripped_fname.rjust(max_fbase_length), fname_ext,
-                             stripped_dest.rjust(max_fbase_length), dest_ext,
-                             output_track.format()))
-
-        if output_track.db is not None:
-            output_track.db.update(encoded, False)
-            for k, v in sorted(output_track.db.changes().items()):
-                print('%s\t%s -> %s' % (k, output_track.db.staged.get(k, None), v))
             if not test:
-                output_track.db.save()
-        print()
+                encoded = open_music_file(dest)
+            else:
+                encoded = Metadata.from_dict({'bitrate': bitrate * 1000,
+                                              'fsize': int(input_track.length * bitrate / 8)})
+
+            if output_track.db is not None:
+                output_track.db.update(encoded, False)
+
+            # Fancy string formatting
+            stripped_fname, fname_ext = os.path.splitext(fname)
+            stripped_dest, dest_ext = os.path.splitext(dest)
+            replace_str = ' (replacing)' if os.path.isfile(dest) else ''
+            max_fbase_length = max(len(stripped_fname), len(stripped_dest))
+
+            print('Transcoded %s%s\n        to %s%s%s with the metadata\n%s' %
+                                (stripped_fname.rjust(max_fbase_length), fname_ext,
+                                 stripped_dest.rjust(max_fbase_length), dest_ext,
+                                 replace_str, output_track.format()))
+
+            if output_track.db is not None:
+                output_track.db.update(encoded, False)
+                for k, v in sorted(output_track.db.changes().items()):
+                    print('%s\t%s -> %s' % (k, output_track.db.staged.get(k, None), v))
+                if not test:
+                    output_track.db.save()
+            print()
 
     for track in unmatched_inputs:
-        print('%s NOT MATCHED' % track.location)
+        print('Input %s NOT MATCHED' % track.location)
+        print()
+
+    for track in unmatched_outputs:
+        print('Output %s NOT MATCHED' % track.title)
         print()
 
     if not test:
         config.DefaultDb().commit()
 
-    print('%d/%d matched' % (len(matched), len(input_files)))
+    print('%d out of %d/%d matched' % (len(matched), len(input_files), len(output_tracks)))
 
 def main():
     parser = argparse.ArgumentParser(description="Transcode a set of music files.")
